llm_service/app/requirements.txt
pydantic
fastapi
uvicorn
google-generativeai
python-dotenv
ollama